data-generator/requirements.txt
kafka-python==2.0.2
Faker==18.11.2